safepp_pytorch/requirements.txt
torch>=2.2
torchvision>=0.17
timm>=1.0.0
pytorch-wavelets>=1.3.0
pandas>=2.0
scikit-learn>=1.4
PyYAML>=6.0
Pillow>=10.0
tqdm>=4.66
